4_RagPractice/requirements.txt
streamlit==1.46.1
langchain==0.3.26
langchain-community==0.0.13
langchain-core==0.3.66
chromadb==1.0.13
transformers==4.53.0
torch==2.7.1
sentence-transformers==4.1.0
python-dotenv==1.1.1
pypdf==5.6.1
langchain-upstage==0.0.2